Test and fix unit_cost
This might prove to be a bad assumption, the unit_cost table works out
the cost of ingredients and intermediate stages at a per kilogram rate,
but the product cost is the cost for a finished product. Where a
finished product is (say) 6 scotch rolls, then the 'unit' is taken to be
a 'batch of six rolls'.

Note too that this 'per_unit' cost is solely the cost of ingredients.
The still to come 'cost_breakdown' view is where we calculate that.