Add a 'rest' column to the recipe table
Where the recipe is an intermediate product, this is the amount of time it
needs to rest before it can be used; if it's a final product, then it's the
bake time.

For sequencing purposes, intermediate products can usually take their rests in
parallel, but bake times are limited by oven space. Not that anything in the
database is doing anything with that information yet.