Constrain recipe_item.ingredient
The ingredient in a recipe item must be an actual basic ingredient or the name
of a recipe. If there's a typo in an ingredient name then this can screw up
the reports we use to decide how much dough to make. This is Not Good. So add
a CHECK constraint to the table to ensure that it doesn't happen. Again.