Fix the rank/ingredient/weight/work_date collision in production_list
It turns out that if multiple orders end up being made from identical weights of
the same intermediate product, the next recursive step only counts one of
those rows. Which means we make less dough than we need. Which can be
something of a catastrophe if we don't end up with enough dough to
fulfil our wholesale orders (not being able to stock the retail shelf
is bad, but less bad than not being able to fulfil our _definite_ sales)

We fix this by adding a 'prd' column to the recursive table, causing
the colliding lines to be distinct again, and thus all counting towards
the total. Which is nice.

We even test the bloody thing properly.